refactor large WiP
add CjElements API (DOM like)
fix CJ-JSON, JSON-CJ mappings
add IJsonValue API (DOM like)

+++ base/src/main/java/com/calpano/graphinout/base/cj/element/ICjGraphMetaProperties.java
@@ -8,15 +8,15 @@ public interface ICjGraphMetaProperties {
     Boolean canonical();
 
     @Nullable
-    Integer nodeCountTotal();
+    Long nodeCountTotal();
 
     @Nullable
-    Integer edgeCountTotal();
+    Long edgeCountTotal();
 
     @Nullable
-    Integer nodeCountInGraph();
+    Long nodeCountInGraph();
 
     @Nullable
-    Integer edgeCountInGraph();
+    Long edgeCountInGraph();
 
 }
